test 코드변경 - 테스트케이스 분리 cars / car / Number - 랜덤값 변경 테스트를 위해 NumberGenerateStrategy 구현클래스 생성 (MovableNumberGenerator , NonMoveableNumberGenerator)
