+++ src/test/java/racingGame/strategy/MovableNumberGenerator.java
@@ -1,0 +1,9 @@
+package racingGame.strategy;
+
+public class MovableNumberGenerator implements NumberGenerateStrategy{
+
+    @Override
+    public int NumberGenerate() {
+        return 4;
+    }
+}
