+++ src/test/java/racingGame/strategy/NonMoveableNumberGenerator.java
@@ -1,0 +1,9 @@
+package racingGame.strategy;
+
+public class NonMoveableNumberGenerator  implements NumberGenerateStrategy{
+
+    @Override
+    public int NumberGenerate() {
+        return 3;
+    }
+}
